+++ src/main/java/org/example/course2024/repository/PriceRepository.java
@@ -5,11 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface PriceRepository extends JpaRepository<Price, Long> {
 
     @Query("SELECT p FROM Price p WHERE p.price BETWEEN :minPrice AND :maxPrice")
-    List<Price> findByPriceRange(@Param("minPrice") double minPrice, @Param("maxPrice") double maxPrice);
+    List<Price> findByPriceRange(@Param("minPrice") BigDecimal minPrice, @Param("maxPrice") BigDecimal maxPrice);
 
 }
